Renamed root dir
