+++ .gitignore
@@ -102,7 +102,7 @@ celerybeat.pid
 *.sage.py
 
 # Environments
-StockDashboardAPI/.env
+stock_dashboard_api/.env
 .venv
 env/
 venv/
