+++ stock_dashboard_api/config.py
@@ -1,7 +1,7 @@
 import os
 from dotenv import load_dotenv
 
-project_folder = os.path.expanduser('~/StockDashboard/StockDashboardAPI')  # adjust as appropriate
+project_folder = os.path.expanduser('~/StockDashboard/stock_dashboard_api')  # adjust as appropriate
 load_dotenv(os.path.join(project_folder, '.env'))
