+++ stock_dashboard_api/routes.py
@@ -1,6 +1,6 @@
-from StockDashboardAPI import app
+from stock_dashboard_api import app
 from flask import render_template, redirect, url_for, request
-from StockDashboardAPI.rabbit.send_email_queue import get_email_queue
+from stock_dashboard_api.rabbit.send_email_queue import get_email_queue
 from .forms import EmailForm
 import pika
 import os
