+++ stock_dashboard_api/run.py
@@ -1,4 +1,4 @@
-from StockDashboardAPI import app
+from stock_dashboard_api import app
 
 if __name__ == '__main__':
     app.run(debug=True)
