+++ stock_dashboard_api/templates/email.html
