+++ stock_dashboard_api/templates/home.html
