+++ stock_dashboard_api/templates/send_email.html
